requirements.txt
astroid==1.6.0
atomicwrites==1.2.1
attrs==18.2.0
backports.functools-lru-cache==1.5
cffi==1.11.5
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
grpcio==1.18.0
grpcio-tools==1.18.0
ipaddr==2.2.0
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
logilab-common==1.4.1
matplotlib==2.2.3
mccabe==0.6.1
more-itertools==5.0.0
networkx==2.1
numpy==1.16.0
pathlib2==2.3.3
pexpect==4.2.1
pluggy==0.8.1
prettytable==0.7.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
py==1.7.0
pybind11==2.2.4
pycparser==2.19
pyflakes==1.6.0
pylint==1.8.3
pyparsing==2.3.1
pytest==4.1.1
python-dateutil==2.7.5
pytz==2018.9
scandir==1.9.0
scapy==2.3.3
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
subprocess32==3.5.3
termcolor==1.1.0
thrift==0.9.2
wcwidth==0.1.7
wrapt==1.9.0
